Simulation v0.1
A model of a sphere and an orbiting body, using Newtonian gravitational forces and a basic Euler method with values for position and velocity in 3D, as well as time passed and altitude above the surface of the sphere. This is the first bit of code that will run with some notable result. 

All constants, as well as starting conditions and initial variables, can be easily changed and are defined in the opening lines.

The sphere is plotted proportional to the size of the body in question, and enough iterations are run for one complete circular orbit around it. I include a break clause in the case the orbit dips below the defined surface of the body around which it orbits. This effectively represents an impact with the ground.

Due to issues with matplotlib in 3D I also create two artificial points equal to the min/max x,y,z values to ensure the 3D space has equal length axis and so is an accurate representation of a physical space. 

Known bugs: 
Note: none of these affect the data and number output of the simulation, only the presentation and formatting of the graphical output.
- Sphere plotted at appear in front of orbit regardless of orbit being in front of/behind sphere. This is due to matplotlib plotting it as two separate layers rather than in true 3D, so it is very difficult to display the orbit so it appears in front or behind the sphere based on its actual location.
- Axis labels still seem uneven despite the axis being the same length. Likely just a formatting issue.